requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
boto3==1.28.71
botocore==1.31.71
cachetools==4.2.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.3.1
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.3
fonttools==4.38.0
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.51.1
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.0.0
imutils==0.5.4
jmespath==1.0.1
joblib==1.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.6.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
numpy==1.21.4
oauthlib==3.2.2
openpyxl==3.1.0
opt-einsum==3.3.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
platformdirs==3.10.0
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
s3transfer==0.7.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sqlparse==0.4.3
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==2.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
virtualenv==20.24.4
Werkzeug==2.3.7
whitenoise==6.6.0
wrapt==1.14.1
xlwt==1.3.0
zipp==3.13.0
